crm-service: fix bug after refactoring code
VHT:
- create API for retrieving VHT history
- sync data from VHT server every 30s, if there are any errors it will sync data after 5 minutes
- in the first time tun server, sync data from 24*30*8 hours to present
- sync data by sdk_call_id can't be used because the corresponding API in VHT server has some problems

Update ticket:
- split func CreateOrUpdateTicket to CreateTicket and UpdateTicket
- sync data VtigerContact
- define role
- delele sqlstore in etop.vn/backend/cmd/supporting/crm-service

Update function:
- ping server vht & vtiger before sync

+++ pkg/services/crm-service/vtiger/client/model.go
@@ -54,8 +54,8 @@ type VtigerServiceTokenResponse struct {
 // VtigerTokenResult properti result of response get token
 type VtigerTokenResult struct {
 	Token      string `json:"token"`
-	ServerTime int32  `json:"serverTime"`
-	ExpireTime int32  `json:"expireTime"`
+	ServerTime int64  `json:"serverTime"`
+	ExpireTime int64  `json:"expireTime"`
 }
 
 // VtigerSessionResponse session response from vtiger
